requirements.txt
argparse==1.2.1
bunch==1.0.1
jsonrpc==1.2
python-jsonrpc==0.6.1
wsgiref==0.1.2
